Content review and updates
Content review @NickCorcoran kindly review the comments on the changes, there are some that need extra context and some information that requires to be updated or removed.

There may be some repetition with certain subjects, please kindly review it as I am not sure if it is necessary. Thank you!